feat: Add Dockerfile and instructions for building and running the syslog server
This commit adds a Dockerfile and instructions for building and running the syslog server. The Dockerfile sets up a basic syslog server that listens on port 514/udp and writes log messages to /var/log/syslog. The instructions in the readme.md file explain how to build the Docker image and run the syslog server container.